Rename README.md to task-1/TO DO LIST APPLICATION
In traditional we use to write in dairy or in notepad or used to write in a piece of paper. But as of the technology has been growing up we have been able to create to do list in a website. You can use a to do list in your home and personal life, or in the workplace.